% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrobel.R
\docType{data}
\name{wrobel_lung}
\alias{wrobel_lung}
\title{wrobel_lung}
\format{
A \link[base]{data.frame}
}
\usage{
wrobel_lung
}
\description{
From \url{http://juliawrobel.com/MI_tutorial/Data/lung.RDA}.
}
\keyword{datasets}
