% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrobel.R
\docType{data}
\name{wrobel_ovarian}
\alias{wrobel_ovarian}
\title{wrobel_ovarian}
\format{
A \link[base]{data.frame}
}
\usage{
wrobel_ovarian
}
\description{
From \url{http://juliawrobel.com/MI_tutorial/Data/ovarian.RDA}.
}
\keyword{datasets}
